Read PCI config space successfully

+++ src/kernel/include/ioport.h
@@ -19,7 +19,7 @@ static inline uint16_t inw(uint16_t port) {
     return ret;
 }
 
-static inline uint16_t inl(uint16_t port) {
+static inline uint32_t inl(uint16_t port) {
     uint32_t ret;
     asm volatile ("inl %1, %0"
                    : "=a"(ret)
